backend/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.25.0
apscheduler>=3.8.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
websocket-client>=1.6.0 